Move db files to specific folder in Network
