+++ apps/network/tests/database/test_role.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.app.users.role import Role
+from src.app.database.role import Role
 from .presets.role import role_metrics
